move detect types to voiceDetect file
